storage permission handler moved to common module
